make sure matomo_filter_incompatible_plugins is available to uninstall script
